project.java
The program lets the user keep a persistent phone book that contains names, email addresses, and phone numbers. The data for the phone book is stored in a file in the user's home directory. The program uses a Tree Map data structure to store the phone directory entries, where the
entries are name/number pairs. When the program is run, it reads the phone book data from the file (if the file exists) and initializes the phone directory. The program then prompts
the user to select an action to perform, which can be looking up a phone number using a name or email, adding or changing a phone number, removing an entry from the phone directory, listing the entire phone directory, or exiting the program. Depending on the user's input, the program carries out the requested action and updates the phone directory data structure and the phone book data file if necessary. The program also handles error conditions, such as invalid user input or errors reading/writing the phone book data file.